feat: 2024 fall store updated
